[FE][Refactor] QA : 마커에 따른 줌 공통 함수로 분리
